requirements.txt
pandas
requests
sqlalchemy
pymysql
python-dotenv
prophet